Refactor vote endpoint
