+++ pages/api/v1/proposals/vote.ts
@@ -2,21 +2,18 @@ import type { UserVote } from '@prisma/client';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
 
-import { onError, onNoMatch, requireSuperApiKey, requireKeys } from 'lib/middleware';
+import { onError, onNoMatch, requireKeys, requireApiKey } from 'lib/middleware';
 import { castProposalVote } from 'lib/public-api/castProposalVote';
-import type { CastProposalVoteRequestBody } from 'lib/public-api/interfaces';
 import { withSessionRoute } from 'lib/session/withSession';
 
 const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
 
 handler
-  // TODO: require either super api key or "regular" api key
-  .use(requireSuperApiKey)
+  .use(requireApiKey)
   .use(
     requireKeys(
       [
         { key: 'proposalId', truthy: true },
-        { key: 'spaceId', truthy: true },
         { key: 'userId', truthy: true },
         { key: 'choice', truthy: true }
       ],
@@ -24,6 +21,71 @@ handler
     )
   )
   .post(castVote);
+
+/**
+ * @example https://github.com/jellydn/next-swagger-doc/blob/main/example/models/organization.ts
+ *
+ * @swagger
+ * components:
+ *  schemas:
+ *    CastProposalVoteRequestBody:
+ *      required:
+ *        - spaceId
+ *        - proposalId
+ *        - userId
+ *        - choice
+ *      type: object
+ *      properties:
+ *        proposalId:
+ *          type: string
+ *          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
+ *        userId:
+ *          type: string
+ *          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
+ *        choice:
+ *          type: string
+ *          example: Abstain
+ */
+export interface CastProposalVoteRequestBody {
+  proposalId: string;
+  userId: string;
+  choice: string;
+}
+
+/**
+ * @example https://github.com/jellydn/next-swagger-doc/blob/main/example/models/organization.ts
+ *
+ * @swagger
+ * components:
+ *  schemas:
+ *    CastProposalVoteResponseBody:
+ *      type: object
+ *      properties:
+ *        voteId:
+ *          type: string
+ *          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
+ *        userId:
+ *          type: string
+ *          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
+ *        choice:
+ *          type: string
+ *          example: Abstain
+ *        createdAt:
+ *          type: string
+ *          format: date-time
+ *          example: 2022-04-04T21:32:38.317Z
+ *        updatedAt:
+ *          type: string
+ *          format: date-time
+ *          example: 2022-04-04T21:32:38.317Z
+ */
+export interface CastProposalVoteResponseBody {
+  voteId: string;
+  userId: string;
+  choice: string;
+  createdAt: string;
+  updatedAt: string;
+}
 
 /**
  * @swagger
